Migrate to the latest badge standard (#11)
Also clean up the intra-doc links to not use `concat`

I'm planning on documenting this pattern for intra-doc links with a doc
include on https://linebender.org, so that we have something to point
towards. For this, I thought I'd better check that the pattern is as
clean as possible, and it turns out that the `concat` is already
implied, so we can just use that.

---------

Co-authored-by: Kaur Kuut <strom@nevermore.ee>